instagram_bot_poc/requirements.txt
instagrapi==2.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
schedule==1.2.0
logging==0.4.9.6
pytest==7.4.0